errors: show the user a helpful message that their token has expired
This PR tries to interpret an authentication error message. The API can
return an error message in the form:

```
{"error":"invalid_token","error_description":"The access token is
invalid","state":"unauthorized"}
```

which is of non-standard type. To handle it properly, we're trying to
unmarshal the error message into the type:

```go
// authErrorResponse represents an error response from the API
type authErrorResponse struct {
	Error            string `json:"error"`
	ErrorDescription string `json:"error_description"`
	State            string `json:"state"`
}
```

And then check whether the `Error` field is of type `invalid_token`,
which is an indication for expired tokens.