Switch to using a keyring for password storage if available
This logic switches to using a keyring like for example keychain on
MacOS if that is available on the platform.

This provides better and more secure secret storage for the token that
gets generated for a user.

If no keyring implementation is available, things keep working like they
work today.